refactor(prometheus): fold the query API into pkg/prometheus and register it with the API server
